Port pulse back to short*.
git-svn-id: https://svn.code.sf.net/p/sidplay-residfp/code/trunk/sidplay-2.0.9@177 75748be7-0387-4326-94c7-f41ce0883d2c

+++ src/audio/pulse/audiodrv.h
@@ -70,13 +70,13 @@ public:  // --------------------------------------------------------- public
     Audio_Pulse();
     ~Audio_Pulse();
 
-    float *open  (AudioConfig &cfg, const char *name);
+    short *open  (AudioConfig &cfg, const char *name);
     void  close ();
-    float *reset ()
+    short *reset ()
     {
         return _sampleBuffer;
     }
-    float *write ();
+    short *write ();
     void  pause () {;}
 };
